clean up 'domain' args, pass principalDomain through to everything
This renames the various "domain" parameters to avoid collision and
emphasize what exactly is meant. Both getWellKnown() and
parseWellKnownBody() now take two separate parameters. "fromDomain" is
the domain that we're currently fetching a .well-known/browserid file
from; when we're recursively following delegations, fromDomain is
different on each loop. "principalDomain" is constant: it is taken from
the principal's email address and remains the same for each loop.

getWellKnown() passes its fromDomain parameter off to its callback.. I'm
not sure why (might be to avoid late-binding confusion in the delgation
case). That argument was renamed to newDomain in the closures passed
into getWellKnown(), since previously it sometimes overlapped with the
"domain" argument being passed in.

This shouldn't change any behavior. It just paves the way for a later
change to add principalDomain as a query arg.