add domain= queryarg to the .well-known fetch
This should allow multiple primaries to delegate control to the same
server, by letting the server know which domain we're interested in.
This is similar to how the HTTP "Host" header enables named-based
virtual hosting. Both the Verifier and the user agent (specifically the
/wsapi/address_info handler) will add this argument. Servers which host
only one key can ignore it.

This might help Bigtent.