Call .onready if 3rd party cookies are disabled.
Prevents users from being immediately logged out if 3rd party cookies are disabled.

Set a cookie in the client before /wsapi/session_context is requested. session_context checks for the existence of this cookie and returns a new field "cookies" which is true if cookies are enabled, false otw. Checking if the cookie can be read on the server avoid the Android 3.2 and 4.0 where cookies can be read by the browser even if they are disabled.

fixes #2308
fixes #2999